+++ server/src-lib/Hasura/Backends/MSSQL/Meta.hs
@@ -41,6 +41,7 @@ loadDBMetadata pool = do
     Left e          -> throw500 $ T.pack $ "error loading sql server database schema: " <> e
     Right sysTables -> pure $ HM.fromList $ map transformTable sysTables
 
+
 --------------------------------------------------------------------------------
 -- Local types
 
@@ -106,17 +107,18 @@ instance FromJSON (SysForeignKeyColumn) where
 
 transformTable :: SysTable -> (TableName, DBTableMetadata 'MSSQL)
 transformTable tableInfo =
-  let schemaName   = ssName $ staJoinedSysSchema tableInfo
-      tableName    = TableName (staName tableInfo) schemaName
-      tableOID     = OID $ staObjectId tableInfo
+  let schemaName = ssName $ staJoinedSysSchema tableInfo
+      tableName  = TableName (staName tableInfo) schemaName
+      tableOID   = OID $ staObjectId tableInfo
       (columns, foreignKeys) = unzip $ transformColumn <$> staJoinedSysColumn tableInfo
+      foreignKeysMetadata = HS.fromList $ map ForeignKeyMetadata $ coalesceKeys $ concat foreignKeys
   in ( tableName
      , DBTableMetadata
        tableOID
        columns
        Nothing  -- no primary key information?
        HS.empty -- no unique constraints?
-       (HS.fromList $ map ForeignKeyMetadata $ HM.elems $ coalesceKeys $ concat foreignKeys)
+       foreignKeysMetadata
        Nothing  -- no views, only tables
        Nothing  -- no description
      )
@@ -135,8 +137,8 @@ transformColumn columnInfo =
       prciDescription = Nothing
       prciType = parseScalarType $ styName $ scJoinedSysType columnInfo
       foreignKeys = scJoinedForeignKeyColumns columnInfo <&> \foreignKeyColumn ->
-        let _fkConstraint    = Constraint () {- FIXME -} $ OID $ sfkcConstraintObjectId foreignKeyColumn
-            -- ^ there's currently no ConstraintName type in our MSSQL code?
+        let _fkConstraint    = Constraint () $ OID $ sfkcConstraintObjectId foreignKeyColumn
+            -- ^ constraints have no name in MSSQL, and are uniquely identified by their OID
             schemaName       = ssName $ sfkcJoinedReferencedSysSchema foreignKeyColumn
             _fkForeignTable  = TableName (sfkcJoinedReferencedTableName foreignKeyColumn) schemaName
             _fkColumnMapping = HM.singleton prciName $ ColumnName $ sfkcJoinedReferencedColumnName foreignKeyColumn
@@ -147,11 +149,11 @@ transformColumn columnInfo =
 --------------------------------------------------------------------------------
 -- Helpers
 
-coalesceKeys :: [ForeignKey 'MSSQL] -> HM.HashMap TableName (ForeignKey 'MSSQL)
-coalesceKeys = foldl' coalesce HM.empty
-  where coalesce mapping fk@(ForeignKey _ tableName _) = HM.insertWith combine tableName fk mapping
-        -- is it ok to assume we can coalesce only on table name?
-        combine oldFK newFK = oldFK { _fkColumnMapping = (HM.union `on` _fkColumnMapping) oldFK newFK }
+coalesceKeys :: [ForeignKey 'MSSQL] -> [ForeignKey 'MSSQL]
+coalesceKeys = HM.elems . foldl' coalesce HM.empty
+  where
+    coalesce mapping fk@(ForeignKey constraint tableName _) = HM.insertWith combine (constraint, tableName) fk mapping
+    combine oldFK newFK = oldFK { _fkColumnMapping = (HM.union `on` _fkColumnMapping) oldFK newFK }
 
 parseScalarType :: Text -> ScalarType
 parseScalarType = \case